feat: 특정 날짜를 기준으로 api 활성화 여부 판단

+++ app/router/letters.py
@@ -13,9 +13,11 @@ from fastapi_utils.camelcase import camel2snake, snake2camel
 from app.authconfig import settings
 from fastapi_cognito import CognitoToken, CognitoAuth, CognitoSettings
 from typing import List
+from app.common.util import Util
 
 router = APIRouter()
 cognito_kr = CognitoAuth(settings=CognitoSettings.from_global_settings(settings), userpool_name="kr")
+util = Util()
 
 @router.get("/users/{user_id}/letters", tags=["letters"], response_model=List, response_model_exclude_none=True)
 async def get_letters(user_id: int, page: int=0, db: Session = Depends(get_db)):
@@ -52,6 +54,9 @@ async def get_letter_detail_by_password(user_id: int, letter_id: int, letter: Le
 
 @router.get("/letters/{letter_id}", tags=["letters"], response_model=LetterDetail, response_model_exclude_none=True)
 async def get_letter_detail_by_jwt(letter_id: int, auth: CognitoToken = Depends(cognito_kr.auth_required), db: Session = Depends(get_db)):    
+    if not util.is_passed_basedate():
+        raise HTTPException(status_code=400, detail="사용할 수 없습니다.")
+    
     db_letter = await LetterRepository.fetch_by_id(db, _id=letter_id)
 
     db_owner = await OwnerRepository.fetch_by_id(db, _id=db_letter.owner_id)
@@ -66,6 +71,9 @@ async def get_letter_detail_by_jwt(letter_id: int, auth: CognitoToken = Depends(
 
 @router.patch("/letters/{letter_id}", tags=["letters"], response_model=LetterDetail, response_model_exclude_none=True)
 async def update_letters(letter: LetterModify, db: Session = Depends(get_db)):
+    if util.is_passed_basedate():
+        raise HTTPException(status_code=400, detail="사용할 수 없습니다.")
+
     db_letter = await LetterRepository.update(db, letter)
     sticker = await StickerRepository.fetch_by_id(db, _id=db_letter.sticker_id)
 
@@ -75,6 +83,9 @@ async def update_letters(letter: LetterModify, db: Session = Depends(get_db)):
 
 @router.delete("/letters/{letter_id}", tags=["letters"])
 async def delete_letters(letter_id: int, db: Session = Depends(get_db)):
+    if util.is_passed_basedate():
+        raise HTTPException(status_code=400, detail="사용할 수 없습니다.")
+
     db_letter = await LetterRepository.fetch_by_id(db, _id=letter_id)
     await LetterRepository.delete(db, db_letter)
 
@@ -82,6 +93,9 @@ async def delete_letters(letter_id: int, db: Session = Depends(get_db)):
 
 @router.post("/users/{user_id}/letters", tags=["letters"], response_model=LetterDetail, response_model_exclude_none=True)
 async def create_letters(user_id: int, letter: Letter, db: Session = Depends(get_db)):
+    if util.is_passed_basedate():
+        raise HTTPException(status_code=400, detail="사용할 수 없습니다.")
+
     owner = await OwnerRepository.fetch_by_id(db, user_id)
     
     setattr(letter, 'owner_id', user_id)
